OWNERS: promote Zev to maintainer
Zev is an active reviewer of this repository and seems to have a
good handle on the needs from a user perspective.  Suggest promoting
him to maintainer to aid in responsiveness to reviews on this repo.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Id00825858f55582d0bb1f72657fe3fcc5b03eed1
(cherry picked from commit a6a4da401acef16c2149e91853a9e0595a7c7fb6)
Signed-off-by: Joseph Liu <kwliu@nuvoton.com>